requirements.txt
asgiref==3.4.1
async-generator==1.10
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cryptography==36.0.0
Django==3.2.10
django-cors-headers==3.10.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
greenlet==1.1.2
h11==0.12.0
idna==3.3
outcome==1.1.0
playwright==1.17.2
pycparser==2.21
pyee==8.2.2
PyJWT==1.7.1
pyOpenSSL==21.0.0
pytz==2021.3
requests==2.26.0
selenium==4.1.0
selenium-stealth==1.0.6
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
TikTokApi==4.1.0
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.0.1
urllib3==1.26.7
websockets==10.1
wsproto==1.0.0
